Temporarily deactivated SCHEMA_VALIDATION test cases: TSM_0350 and TSM_0351
