+++ test/pytest/pytestfiles/test_02_VARIANT_HANDLING_BADCASE.py
@@ -18,7 +18,7 @@
 #
 # XC-CT/ECA3-Queckenstedt
 #
-# 31.05.2023 - 15:27:30
+# 06.06.2023 - 15:28:28
 #
 # --------------------------------------------------------------------------------------------------------------
