+++ test/pytest/pytestfiles/test_08_ROBOT_CODE_BADCASE.py
@@ -14,11 +14,11 @@
 #  limitations under the License.
 # --------------------------------------------------------------------------------------------------------------
 #
-# test_09_ROBOT_CODE_BADCASE.py
+# test_08_ROBOT_CODE_BADCASE.py
 #
 # XC-CT/ECA3-Queckenstedt
 #
-# 31.05.2023 - 15:27:30
+# 06.06.2023 - 15:28:29
 #
 # --------------------------------------------------------------------------------------------------------------
