+++ test/testconfig/TestConfig.py
@@ -22,7 +22,7 @@
 #
 # --------------------------------------------------------------------------------------------------------------
 #
-# 02.06.2023
+# 06.06.2023
 #
 # --------------------------------------------------------------------------------------------------------------
 #
@@ -31,6 +31,10 @@
 # Because of the return value computation in RobotFramework AIO is currently under discussion, in all BADCASE tests,
 # that cause a return value != 0, the 'EXPECTEDRETURN' is set to None.
 # This deactivates the return value check.
+#
+# 06.06.2023
+# TSM_0350 and TSM_0351 temporarily deactivated.
+# Reason: Access right problems under Linux need to be clarified.
 #
 # --------------------------------------------------------------------------------------------------------------
 
@@ -654,38 +658,45 @@ listofdictUsecases.append(dictUsecase)
 del dictUsecase
 # --------------------------------------------------------------------------------------------------------------
 # --------------------------------------------------------------------------------------------------------------
-dictUsecase = {}
-dictUsecase['TESTID']           = "TSM_0350"
-dictUsecase['DESCRIPTION']      = "Schema file for JSON configuration files is not available"
-dictUsecase['EXPECTATION']      = "Test is not executed; error message; test result is UNKNOWN"
-dictUsecase['SECTION']          = "SCHEMA_VALIDATION"
-dictUsecase['SUBSECTION']       = "BADCASE"
-dictUsecase['COMMENT']          = "Single file execution"
-dictUsecase['HINT']             = "Temporary modification of installed schema file"
-dictUsecase['TESTFILENAME']     = "tsm-testfile-02.robot" # (with variant configuration)
-dictUsecase['TESTFOLDERNAME']   = None
-dictUsecase['ADDITIONALPARAMS'] = "--variable variant:\"variant1\""
-dictUsecase['PRESTEP']          = "ConfigSchemaFile_Remove"
-dictUsecase['POSTSTEP']         = "ConfigSchemaFile_Restore"
-dictUsecase['EXPECTEDRETURN']   = None # 256
-listofdictUsecases.append(dictUsecase)
-del dictUsecase
+
 # --------------------------------------------------------------------------------------------------------------
-dictUsecase = {}
-dictUsecase['TESTID']           = "TSM_0351"
-dictUsecase['DESCRIPTION']      = "Schema file for JSON configuration files is invalid because of a syntax error"
-dictUsecase['EXPECTATION']      = "Test is not executed; error message; test result is UNKNOWN"
-dictUsecase['SECTION']          = "SCHEMA_VALIDATION"
-dictUsecase['SUBSECTION']       = "BADCASE"
-dictUsecase['COMMENT']          = "Single file execution"
-dictUsecase['TESTFILENAME']     = "tsm-testfile-02.robot" # (with variant configuration)
-dictUsecase['TESTFOLDERNAME']   = None
-dictUsecase['ADDITIONALPARAMS'] = "--variable variant:\"variant1\""
-dictUsecase['PRESTEP']          = "ConfigSchemaFile_MakeInvalid"
-dictUsecase['POSTSTEP']         = "ConfigSchemaFile_Restore"
-dictUsecase['EXPECTEDRETURN']   = None # 256
-listofdictUsecases.append(dictUsecase)
-del dictUsecase
+# 06.06.2023
+# TSM_0350 and TSM_0351 temporarily deactivated.
+# Reason: Access right problems under Linux need to be clarified.
+# --------------------------------------------------------------------------------------------------------------
+
+# # dictUsecase = {}
+# # dictUsecase['TESTID']           = "TSM_0350"
+# # dictUsecase['DESCRIPTION']      = "Schema file for JSON configuration files is not available"
+# # dictUsecase['EXPECTATION']      = "Test is not executed; error message; test result is UNKNOWN"
+# # dictUsecase['SECTION']          = "SCHEMA_VALIDATION"
+# # dictUsecase['SUBSECTION']       = "BADCASE"
+# # dictUsecase['COMMENT']          = "Single file execution"
+# # dictUsecase['HINT']             = "Temporary modification of installed schema file"
+# # dictUsecase['TESTFILENAME']     = "tsm-testfile-02.robot" # (with variant configuration)
+# # dictUsecase['TESTFOLDERNAME']   = None
+# # dictUsecase['ADDITIONALPARAMS'] = "--variable variant:\"variant1\""
+# # dictUsecase['PRESTEP']          = "ConfigSchemaFile_Remove"
+# # dictUsecase['POSTSTEP']         = "ConfigSchemaFile_Restore"
+# # dictUsecase['EXPECTEDRETURN']   = None # 256
+# # listofdictUsecases.append(dictUsecase)
+# # del dictUsecase
+# # # --------------------------------------------------------------------------------------------------------------
+# # dictUsecase = {}
+# # dictUsecase['TESTID']           = "TSM_0351"
+# # dictUsecase['DESCRIPTION']      = "Schema file for JSON configuration files is invalid because of a syntax error"
+# # dictUsecase['EXPECTATION']      = "Test is not executed; error message; test result is UNKNOWN"
+# # dictUsecase['SECTION']          = "SCHEMA_VALIDATION"
+# # dictUsecase['SUBSECTION']       = "BADCASE"
+# # dictUsecase['COMMENT']          = "Single file execution"
+# # dictUsecase['TESTFILENAME']     = "tsm-testfile-02.robot" # (with variant configuration)
+# # dictUsecase['TESTFOLDERNAME']   = None
+# # dictUsecase['ADDITIONALPARAMS'] = "--variable variant:\"variant1\""
+# # dictUsecase['PRESTEP']          = "ConfigSchemaFile_MakeInvalid"
+# # dictUsecase['POSTSTEP']         = "ConfigSchemaFile_Restore"
+# # dictUsecase['EXPECTEDRETURN']   = None # 256
+# # listofdictUsecases.append(dictUsecase)
+# # del dictUsecase
 # --------------------------------------------------------------------------------------------------------------
 # --------------------------------------------------------------------------------------------------------------
 dictUsecase = {}
